+++ ReleaseNotes.md
@@ -179,3 +179,4 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 #### Features
 * Added `XmlnsDefinition` `https://spiegelp.github.io/MaterialDesignExtensions/winfx/xaml`
 * Consider `ResizeMode` for the window caption buttons
+* `ResizeGrip` in `MaterialWindow`